Refactorizacion de codigo duplicado en tangram y mensajes en consola del navegador
