+++ src/services/crudService.js
@@ -4,8 +4,6 @@ export const getTherapistServ = async (cedulaT) => {
 	try {
 		const response = await fetch(API_THERAPIST + '/' + cedulaT)
 		const respData = await response.json()
-			
-		console.log(respData)
 		return respData
 	} catch (error) {
 		console.error('Error login user:', error)
@@ -17,8 +15,6 @@ export const getPatientsServ = async (cedulaT) => {
 	try {
 		const response = await fetch(API_PATIENT + '/' + cedulaT)
 		const respData = await response.json()
-
-		console.log(respData)
 		return respData
 	} catch (error) {
 		console.error('Error login user:', error)
@@ -30,8 +26,6 @@ export const getPatientSessionsServ = async (cedulaP) => {
 	try {
 		const response = await fetch(API_SESSION + '/' + cedulaP)
 		const respData = await response.json()
-
-		console.log(respData)
 		return respData
 	} catch (error) {
 		console.error('Error login user:', error)
@@ -43,9 +37,6 @@ export const getHistoryPatientTransferServ = async (cedulaP) => {
 	try {
 		const response = await fetch(API_HISTORY_TRANSFER + '/' + cedulaP)
 		const respData = await response.json()
-
-		//TODO: JEFF - ver la respuesta
-		console.log(respData)
 		return [
 			{
 				date: '2025-12-10',
@@ -75,7 +66,6 @@ export const getHistoryPatientTransferServ = async (cedulaP) => {
 //TODO:implementar
 export const registerSessionServ = async (session) => {
 	try {
-		console.log("Datos enviados al backend desde crud-service", session);
 		const response = await fetch(API_SESSION + '/register', {
 			method: 'POST',
 			headers: {
@@ -83,17 +73,16 @@ export const registerSessionServ = async (session) => {
 			},
 			body: JSON.stringify(session),
 		})
-		const respData = await response.json();
-    	console.log("Respuesta del backend:", respData);
+		const respData = await response.json()
 	} catch (error) {
 		console.error('Error register session', error)
 		throw error
 	}
 }
+
 //TODO:implementar
 export const registerPatientService = async (patient) => {
 	try {
-		console.log("Datos enviados al backend desde crud-service", patient);
 		const response = await fetch(API_PATIENT + '/register', {
 			method: 'POST',
 			headers: {
@@ -102,15 +91,14 @@ export const registerPatientService = async (patient) => {
 			body: JSON.stringify(patient),
 		})
 
-		 if (!response.ok) {
-            const errorData = await response.json();
-            console.error("Error en la solicitud POST:", errorData);
-            throw new Error(errorData.msg || "Error desconocido");
-        }
+		if (!response.ok) {
+			const errorData = await response.json()
+			console.error("Error en la solicitud POST:", errorData)
+			throw new Error(errorData.msg || "Error desconocido")
+		}
 
-		const respData = await response.json();
-        console.log("Respuesta del backend:", respData);
-        return respData;
+		const respData = await response.json()
+		return respData
 	} catch (error) {
 		console.error('Error login user:', error)
 		throw error
@@ -120,8 +108,6 @@ export const registerPatientService = async (patient) => {
 //TODO:implementar
 export const transferPatientService = async (transfer) => {
 	try {
-		
-		console.log("Datos enviados al backend desde crud-service", transfer);
 		const response = await fetch(API_TRANSFER + '/register', {
 			method: 'POST',
 			headers: {
@@ -130,9 +116,8 @@ export const transferPatientService = async (transfer) => {
 			body: JSON.stringify(transfer),
 		})
 
-		 const respData = await response.json()
-		 console.log("Respuesta del backend:", respData);
-		 return respData
+		const respData = await response.json()
+		return respData
 	} catch (error) {
 		console.error('Error login user:', error)
 		throw error
